MachineSSAUpdater: insert IMPLICIT_DEF at top of basic block
When getValueInMiddleOfBlock happens to be called for a basic block
that has no incoming value at all, an IMPLICIT_DEF is inserted in that
block via GetValueAtEndOfBlockInternal. This IMPLICIT_DEF must be at
the top of its basic block or it will likely not reach the use that
the caller intends to insert.

TODO: Add a MIR test for SILowerI1Copies

Issue: https://github.com/GPUOpen-Drivers/llpc/issues/204