Analysis: Add a GenericCycleInfo analysis
This analysis allows algorithms to capture all loops, including irreducible
loops, in a structured way that meshes nicely with the dominator tree:
every cycle is defined by a set of siblings ("headers") in the dominator tree
and a set of other blocks and cycles, each of which are dominated by one of
the "headers.

See the definition of a "cycle" in the header file and the IR test cases for
more detail.