IR: Add convergence control operand bundle and intrinsics
See ConvergentOperations.rst for the details.

This replaces the proposal from https://reviews.llvm.org/D68994

This patch adds the operand bundle and intrinsics themselves, as well as
the LangRef documentation describing the semantics of controlled
convergent operations. Follow-up patches will adjust existing passes to
comply with those changes, as well as provide new functionality on top
of this mechanism.

v8:
- tighten the static rules about cycles: there was a gap in the exact
  phrasing if two loop heart intrinsics in a cycle use _different_
  convergence tokens
- add verifier checks and corresponding tests for the static rules

Change-Id: I045c6bc864c4dc5fb0a23b0279e30fac06c5b974

Differential Revision: https://reviews.llvm.org/D85603