(WIP) AMDGPU: New control flow lowering
Perform SelectionDAG ISel on the original, thread-level control flow
graph and transform to a wave-level control flow graph immediately
afterwards.

This allows us to get rid of the old structurization and replace it
with a reconverging transform that can preserve more uniform branches.
Furthermore, it is a step towards preserving the thread-level CFG,
which will allow us to compute tighter live intervals for VGPRs
(outside of WWM).

The new control flow lowering is enabled with
-amdgpu-use-new-control-flow=1.

v3:
- Fix bug with infinite loops
- Fix: Correctly rotate the current node to be the primary
  successor after rerouting during the reconvergence transform

v5:
- use is_contained instead of `llvm::find(range, ...) != std::end(range)` pattern
- additional cleanups

TODO:
- tests