Add llvm::llvm_fast_shutdown()
LLD uses an unclean process exit to shut the process down faster, but
still wants the benefit of some side-effects that have traditionally
been invoked via llvm_shutdown() and ManagedStatic destruction.

The new llvm_fast_shutdown() function provides a more explicit alternative
solution that will still work when ManagedStatic is removed. Incidentally,
it also reduces the amount of cleanup code that is run even further.

This change also removes the use of ManagedStatic in the affected
components.

Note that the StatLock is moved into the StatisticInfo class. This is a
robust solution to guarantee that the StatLock destructor doesn't run
too early.

v2:
- add documentation to the ProgrammersManual
- flush output streams

v3:
- add some notes about data race safety
- add the DebugCounters to the fast shutdown path

Differential Revision: https://reviews.llvm.org/D129119