[AMDGPU] Combine BFI instructions.
When translating multiple bitfieldInserts where one
bitfieldInsert is the base for another one, the current
mechanism of creating XOR, AND, XOR sequences in the
frontend and having them lowered in ISel is not sufficient,
as the information about the bitmasks is lost during InstCombine.

When creating the canonical bitfieldInsert pattern directly in the
frontend, the constants will still be merged by SimplifyDemandedBits,
leading to a simplified pattern, however, the original sequence can
be reconstructed and the nested bitfieldInsert can be generated.