TableGen: include in libLLVM-*.so
Use of TableGen has grown over the years, so the claim that it is for
internal use only carries less and less water.

The main motivation for adding it to the dylib is to cleanup linking of
tablegen tools outside of llvm itself.

See: https://discourse.llvm.org/t/rfc-cleaning-up-how-we-link-tablegen-tools/66678
Differential Revision: https://reviews.llvm.org/D138277