[IR] Add TargetExtTypeClass and (de)serialize target type info
This allows for extensibility in how the target type info is obtained.

It is a small step towards "Dialects in LLVM".  Users of LLVM can now define
their own "target" extension types (now a somewhat misleading name) with
custom layout and properties, to represent high-level concepts that are
lowered away before code reaches the actual backend.

Target types are still identified by their name and type/int parameters.
However, users can register TargetExtTypeClasses that allow computing
type info and verifying the validity of target types.

During normal operations, TargetExtType::get will query the registered
type class, if any, for the necessary type info.

The type info is serialized when writing bitcode or IR assembly, so that
it can be read back in, and the resulting IR can be processed reasonably
by tools that are not aware of target type classes, such as opt and
bugpoint.

The serialization in bitcode may seem wasteful, but the point is to
future-proof the format by adding self-describing features. There is
redundancy when encoding multiple target types, but this will be
addressed by a follow-up change that introduces an appropriate abbreviation.

v2:
- explicitly list properties in LangRef
- add negative parser/assembler tests
- use tuple binding syntax

v3:
- fix regression: the default layout type is void

Differential Revision: https://reviews.llvm.org/D147697